Main code
In this study, automatic recognition of fish species is discussed using deep learning-based object detection methods such as YOLOv8 and experimentally developed YOLOv11 models. Within the scope of the project, a large visual dataset including different aquaculture species was created; especially images belonging to the Diplodus genus were emphasized and the performance of the model on this genus was tested in detail. Normalization of images and noise reduction techniques were applied in the data pre-processing stage, while transformation, scaling and color variations were applied in the data augmentation stage. The performance of the models was measured with accuracy, precision and recall metrics; although both models varied according to the species covering the dataset, they generally showed success above 85%, but it was observed that YOLOv11 achieved a higher success rate of up to 88%. In addition, a user-friendly interface was developed to increase the practical applicability of the study and real-time recognition results were visualized. The findings reveal the potential of automation-based systems in the aquaculture sector to increase operational efficiency by reducing expert dependency.